+++ vitest.config.ts
@@ -1,6 +1,9 @@
+import { defineConfig } from 'vitest/config'
 import react from '@vitejs/plugin-react'
-import { defineConfig } from 'vite'
 
 export default defineConfig({
   plugins: [react()],
+  test: {
+    environment: 'node',
+  },
 })